+++ flowmachine/flowmachine/core/server/query_schemas/random_sample.py
@@ -27,25 +27,13 @@ class BaseRandomSampleSchema(Schema):
             )
 
 
-class SystemRowsRandomSampleSchema(BaseRandomSampleSchema):
-    # We must define the sampling_method field here for it to appear in the API spec.
-    # This field is removed by RandomSampleSchema before passing on to this schema,
-    # so the sampling_method parameter is never received here and is not included in the
-    # params passed to make_random_sampler.
-    sampling_method = fields.String(validate=OneOf(["system_rows"]))
-
-    @post_load
-    def make_random_sampler(self, params, **kwargs):
-        return RandomSampler(sampling_method="system_rows", **params)
-
-
 class SystemRandomSampleSchema(BaseRandomSampleSchema):
     # We must define the sampling_method field here for it to appear in the API spec.
     # This field is removed by RandomSampleSchema before passing on to this schema,
     # so the sampling_method parameter is never received here and is not included in the
     # params passed to make_random_sampler.
     sampling_method = fields.String(validate=OneOf(["system"]))
-    seed = fields.Float()
+    seed = fields.Float(required=True)
 
     @post_load
     def make_random_sampler(self, params, **kwargs):
@@ -58,7 +46,7 @@ class BernoulliRandomSampleSchema(BaseRandomSampleSchema):
     # so the sampling_method parameter is never received here and is not included in the
     # params passed to make_random_sampler.
     sampling_method = fields.String(validate=OneOf(["bernoulli"]))
-    seed = fields.Float()
+    seed = fields.Float(required=True)
 
     @post_load
     def make_random_sampler(self, params, **kwargs):
@@ -71,7 +59,7 @@ class RandomIDsRandomSampleSchema(BaseRandomSampleSchema):
     # so the sampling_method parameter is never received here and is not included in the
     # params passed to make_random_sampler.
     sampling_method = fields.String(validate=OneOf(["random_ids"]))
-    seed = fields.Float(validate=Range(-1.0, 1.0))
+    seed = fields.Float(validate=Range(-1.0, 1.0), required=True)
 
     @post_load
     def make_random_sampler(self, params, **kwargs):
@@ -80,7 +68,7 @@ class RandomIDsRandomSampleSchema(BaseRandomSampleSchema):
 
 class RandomSampler:
     def __init__(
-        self, *, sampling_method, size=None, fraction=None, estimate_count, seed=None
+        self, *, sampling_method, estimate_count, seed, size=None, fraction=None
     ):
         # Note: all input parameters need to be defined as attributes on `self`
         # so that marshmallow can serialise the object correctly.
@@ -88,8 +76,7 @@ class RandomSampler:
         self.size = size
         self.fraction = fraction
         self.estimate_count = estimate_count
-        if sampling_method != "system_rows":
-            self.seed = seed
+        self.seed = seed
 
     def make_random_sample_object(self, query):
         """
@@ -104,21 +91,18 @@ class RandomSampler:
         -------
         Random
         """
-        sample_params = {
-            "sampling_method": self.sampling_method,
-            "size": self.size,
-            "fraction": self.fraction,
-            "estimate_count": self.estimate_count,
-        }
-        if self.sampling_method != "system_rows":
-            sample_params["seed"] = self.seed
-        return query.random_sample(**sample_params)
+        return query.random_sample(
+            sampling_method=self.sampling_method,
+            size=self.size,
+            fraction=self.fraction,
+            estimate_count=self.estimate_count,
+            seed=self.seed,
+        )
 
 
 class RandomSampleSchema(OneOfSchema):
     type_field = "sampling_method"
     type_schemas = {
-        "system_rows": SystemRowsRandomSampleSchema,
         "system": SystemRandomSampleSchema,
         "bernoulli": BernoulliRandomSampleSchema,
         "random_ids": RandomIDsRandomSampleSchema,